login ve logout düzeltildi

+++ ReactNative/src/containers/menu/settings/changePassword/changePassword.container.tsx
@@ -2,10 +2,9 @@ import React from 'react';
 import { StackActions, NavigationActions } from 'react-navigation';
 import { NavigationStackScreenProps } from 'react-navigation-stack';
 import { ChangePassword, ChangePasswordFormData } from './changePassword.component';
-import firebase from 'firebase';
-import { Alert, AsyncStorage } from 'react-native';
+import { AsyncStorage } from 'react-native';
 import SecureStorage from '@src/core/utils/secure.store';
-import { log } from '@src/core/navigation/util';
+import axios from "axios"
 
 interface State {
   formData: ChangePasswordFormData | undefined;
@@ -26,27 +25,22 @@ export class ChangePasswordContainer extends React.Component<NavigationStackScre
 
     let thizz = this;
     const { navigation } = this.props;
-    var user = firebase.auth().currentUser;
-    var credential = firebase.auth.EmailAuthProvider.credential(
-      user.email,
-      thizz.state.formData.oldPassword
-    );
 
     showLoader(true, "Şifre Değiştiriliyor...")
-    user.reauthenticateWithCredential(credential).then(function () {
-      user.updatePassword(thizz.state.formData.newPassword).then(function () {
-        Alert.alert('Başarı!', 'Şifreniz başarıyla değiştirildi.', [{
-          text: 'Tamam', onPress: () => { console.log('OK Pressed') }
-        }]);
-        showLoader(false, "")
-        thizz.logout(showLoader);
-      });
-    }).catch(function (error) {
-      log({ 'message': 'Girilen şifre yanlıştır lütfen tekrar deneyiniz.' }, 'onChangePasswordButtonPress', false, showLoader);
-      Alert.alert('Hata!', 'Girilen şifre yanlıştır lütfen tekrar deneyiniz.', [{
-        text: 'Tamam', onPress: () => showLoader(false, "")
-      }]);
-    });
+    // user.reauthenticateWithCredential(credential).then(function () {
+    //   user.updatePassword(thizz.state.formData.newPassword).then(function () {
+    //     Alert.alert('Başarı!', 'Şifreniz başarıyla değiştirildi.', [{
+    //       text: 'Tamam', onPress: () => { console.log('OK Pressed') }
+    //     }]);
+    //     showLoader(false, "")
+    //     thizz.logout(showLoader);
+    //   });
+    // }).catch(function (error) {
+    //   log({ 'message': 'Girilen şifre yanlıştır lütfen tekrar deneyiniz.' }, 'onChangePasswordButtonPress', false, showLoader);
+    //   Alert.alert('Hata!', 'Girilen şifre yanlıştır lütfen tekrar deneyiniz.', [{
+    //     text: 'Tamam', onPress: () => showLoader(false, "")
+    //   }]);
+    // });
   };
 
   /**
@@ -56,30 +50,49 @@ export class ChangePasswordContainer extends React.Component<NavigationStackScre
    */
   private logout = (showLoader) => {
     const { navigation } = this.props;
-    showLoader(true, "Çıkış Yapılıyor...");
-    log({ 'message': 'Password change succesful' }, 'logout', true, showLoader);
-    firebase.auth().signOut().then(function () {
-      SecureStorage.deleteData("loginInfo");
-      AsyncStorage.clear();
-      const resetAction = StackActions.reset({
-        index: 0,
-        actions: [NavigationActions.navigate({
-          routeName: 'Auth'
-        })]
-      });
-      showLoader(false, "")
-      navigation.dispatch(resetAction)
-    }).catch(function (error) {
-      const resetAction = StackActions.reset({
-        index: 0,
-        actions: [NavigationActions.navigate({
-          routeName: 'Auth'
-        })]
-      });
-      navigation.dispatch(resetAction)
-      showLoader(false, "")
-      log(error, 'logout', false, showLoader);
+
+    AsyncStorage.getItem('userInfo').then(result => {
+      if (result) {
+        const user = JSON.parse(result);
+        var bodyFormData = new FormData();
+        console.log(user.token)
+        bodyFormData.append('token', user.token);
+
+        showLoader(true, "Çıkış Yapılıyor...");
+        axios({
+          method: 'post',
+          url: 'http://mricheck.calgan.engineer/logout.php',
+          data: bodyFormData,
+          headers: { 'Content-Type': 'multipart/form-data' }
+        }).then(function (response) {
+          //handle success
+          console.log(response.data);
+          SecureStorage.deleteData("loginInfo");
+          SecureStorage.deleteData("userInfo");
+          AsyncStorage.clear();
+          const resetAction = StackActions.reset({
+            index: 0,
+            actions: [NavigationActions.navigate({
+              routeName: 'Auth'
+            })]
+          });
+          showLoader(false, "");
+          navigation.dispatch(resetAction)
+        }).catch(function (response) {
+          //handle error
+          console.log(response);
+          const resetAction = StackActions.reset({
+            index: 0,
+            actions: [NavigationActions.navigate({
+              routeName: 'Auth'
+            })]
+          });
+          navigation.dispatch(resetAction)
+          showLoader(false, "");
+        });
+      }
     });
+
   };
 
   public render(): React.ReactNode {
